Moves user-facing component rendering to DisplayNews component Moves admin-related components to subfolder
